chore: move useCowShedSignature from apps/withdraw-pool to packages/utils
