+++ packages/utils/cowShed/useHookDeadline.ts
@@ -1,8 +1,14 @@
-import { useIFrameContext } from "#/context/iframe";
+import { SupportedChainId } from "@cowprotocol/cow-sdk";
+import type { HookDappContext } from "@cowprotocol/hook-dapp-lib";
 import { useMemo } from "react";
+import { Address } from "viem";
+import type { HookDappContextAdjusted } from "./types";
 
-export function useHookDeadline() {
-  const { context } = useIFrameContext();
+export function useHookDeadline({
+  context,
+}: {
+  context: HookDappContextAdjusted | undefined;
+}) {
   return useMemo(() => {
     return BigInt(
       context?.orderParams?.validTo || generateTimestampOnNextHour()
