+++ packages/utils/index.ts
@@ -2,6 +2,7 @@ export * from "./cowTokenList";
 export * from "./addressUtils";
 export * from "./chainInfo";
 export * from "./cowExplorer";
+export * from "./cowShed";
 
 export { default as parseFixed } from "./parseFixed";
 
@@ -176,8 +177,8 @@ const networksNamesOnBalancer = [
 
 export const networksOnBalancer = Object.fromEntries(
   Object.entries(networkIdEnumMap).filter(([key]) =>
-    networksNamesOnBalancer.includes(key),
-  ),
+    networksNamesOnBalancer.includes(key)
+  )
 );
 
 export function networkFor(key?: string | number) {
@@ -197,7 +198,7 @@ export function networkIdFor(name?: string) {
 
 export function unsafeNetworkIdFor(name: string) {
   return Object.keys(networkIdEnumMap).find(
-    (key) => networkIdEnumMap[key as keyof typeof networkIdEnumMap] === name,
+    (key) => networkIdEnumMap[key as keyof typeof networkIdEnumMap] === name
   );
 }
